pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nocasim"
version = "0.1.0"
description = "Norovirus Capture Sequencing Simulator"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "biopython>=1.81",
    "numpy>=1.24",
    "scipy>=1.10",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[tool.setuptools.packages.find]
include = ["nocasim*"]

[project.scripts]
nocasim = "nocasim.cli:main"
